src/backend/requirements.txt
django
djangorestframework
django-cors-headers
django-polymorphic
django-rest-polymorphic
dateutils
psycopg2-binary